refactor(ui): Extract runSequentialPromises types into a file
